Add inverse lookup capability using Thermochimica Closes #25453
