+++ modules/chemical_reactions/test/tests/thermochimica/MoRu_manual_element.i
@@ -1,0 +1,59 @@
+[Mesh]
+  [gen]
+    type = GeneratedMeshGenerator
+    dim = 2
+    nx = 10
+    ny = 1
+  []
+[]
+
+[GlobalParams]
+  elements = 'Mo Ru'
+  output_phases = 'BCCN HCPN'
+  output_species = 'BCCN:Mo HCPN:Mo BCCN:Ru HCPN:Ru'
+  output_element_potentials = 'mu:Mo mu:Ru'
+  output_vapor_pressures = 'vp:gas_ideal:Mo'
+  output_element_phases = 'ep:BCCN:Mo'
+  make_element_aux_variables = false
+[]
+
+[AuxVariables]
+  [Mo]
+  []
+  [Ru]
+  []
+[]
+
+[ChemicalComposition]
+  thermofile = Kaye_NobleMetals.dat
+  tunit = K
+  punit = atm
+  munit = moles
+  temperature = 2250
+  output_species_unit = mole_fraction
+[]
+
+[ICs]
+  [Mo]
+    type = FunctionIC
+    variable = Mo
+    function = '0.8*(1-x)+4.3*x'
+  []
+  [Ru]
+    type = FunctionIC
+    variable = Ru
+    function = '0.2*(1-x)+4.5*x'
+  []
+[]
+
+[Problem]
+  solve = false
+[]
+
+[Executioner]
+  type = Steady
+[]
+
+[Outputs]
+  exodus = true
+[]
