+++ modules/chemical_reactions/test/tests/thermochimica/gold/MoRu_single_point_out.csv
@@ -1,0 +1,3 @@
+time,Mo,Ru,mu:Mo
+0,0,0,0
+1,0.5,1,-157452.85733323
